requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
audioread==2.1.8
autopep8==1.4.4
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
isort==4.3.21
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
librosa==0.6.3
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.0.3
mccabe==0.6.1
numba==0.44.1
numpy==1.16.4
Pillow==6.1.0
pkg-resources==0.0.0
protobuf==3.8.0
pycodestyle==2.5.0
pydot-ng==2.0.0
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1.1
resampy==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tb-nightly==1.14.0a20190614
tensorboard==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019062701
tf-nightly==1.15.0.dev20190627
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
